threatexchangebackend/requirements.txt
asgiref==3.4.1
autopep8==1.5.4
cachetools==4.1.1
certifi==2020.4.5.1
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.0.7
django-cors-headers==3.5.0
django-extensions==3.1.3
django-heroku==0.3.1
django-rest-swagger==2.2.0
django-shell-plus==1.1.7
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
drf-yasg==1.20.0
facebook-sdk==3.1.0
future==0.18.2
gunicorn==20.0.4
httplib2==0.18.1
idna==2.9
inflection==0.4.0
ipython==1.2.1
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.4
protobuf==3.13.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
PyJWT==2.3.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-twitter==3.5
python3-openid==3.2.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.6
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.5
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.3.1
stripe==2.61.0
text-unidecode==1.3
toml==0.10.1
uritemplate==3.0.1
urllib3==1.25.9
whitenoise==5.3.0
